chore: renames components to better align with their purpoase
